+++ examples/HHTut/HHTut_run.py
@@ -1,9 +1,3 @@
-import sys
-
-if '-nogui' in sys.argv:
-    import netpyne
-    netpyne.__gui__ = False
-    
 import HHTut  # import parameters file 
 from netpyne import sim  # import netpyne sim module