+++ CHANGES.md
@@ -1,16 +1,28 @@
 # Version 0.6.4
 
+- Added analysis.nTE() function to calculate normalized transfer entropy (issue #42)
+
+- Added analysis.granger() function to calculate and plot Spectral Granger Causality (issue #42)
+
+- Added analysis.plotRatePSD() to plot firing rate power spectral density 
+
+- Added analysis.plotShape() to plot 3D morphology of cell and synapse locations
+
+- Added option to fix soma depth for subcellular distributions based on 1d and 2d maps
+
 - Added support for gap junction (electrical synapse) connections
 
-- Added function plotRatePSD to plot firing rate power spectral density 
+- Improved import of multicompartmental cells from NeuroML2
 
-- Added plotShape to plot 3D morphology of cell and synapse locations
+- Added option to turn off graphics by adding -nogui from command line
 
 - Added option to skip batch sims if output file already exists
 
 - Added option to overlay pop labels and show avg rates to plotRaster() (issue #111)
 
 - All section now include argument cell=self to differentiate them 
+
+- Fixed travis script so checks all tuts and examples after each commit (issue #37)
 
 - Fixed bug positioning cells with 3d geom